refactor: move mapping to own file
